warning: implicit conversion from 'std::numeric_limits<unsigned long>::type' (aka 'unsigned long') to 'unsigned int' changes value from 18446744073709551615 to 4294967295 [-Wconstant-conversion]

+++ OgreMain/src/OgreScriptTranslator.cpp
@@ -8460,7 +8460,7 @@ namespace Ogre{
                 }
 
                 shadowMapIdx = StringConverter::parseUnsignedInt( targetPassName,
-                                                                  std::numeric_limits<size_t>::max() );
+                                                                  std::numeric_limits<unsigned>::max() );
 
                 if( shadowMapIdx >= nodeDef->getNumShadowTextureDefinitions() )
                 {
